testing y logic

+++ src/medicos/medicos.service.ts
@@ -15,6 +15,9 @@ export class MedicosService {
 
   async create(createMedicoDto: CreateMedicoDto){
     const {nombre, especialidad, telefono} = createMedicoDto;
+    if (!nombre || !especialidad ){
+        throw new BusinessLogicException('nombre or especialidad must have a valid value', BusinessError.PRECONDITION_FAILED)
+    }
     return await this.medicoRepository.save({nombre, especialidad, telefono})
   }
 
@@ -24,15 +27,19 @@ export class MedicosService {
 
 
   async findOne(id: string) : Promise<MedicoEntity> {
-    const medico:MedicoEntity = await this.medicoRepository.findOne({where: {id}});
+    const medico:MedicoEntity = await this.medicoRepository.findOne({where: {id}, relations: ["pacientes"]});
     if (!medico){
-        throw new BusinessLogicException(`The medic with id ${id} doesn't exists`, BusinessError.NOT_FOUND)
+        throw new BusinessLogicException(`The doctor with id ${id} doesn't exists`, BusinessError.NOT_FOUND)
     }
     return medico;
   }
 
   async delete(id: string) {
     const persistedMedico = await this.findOne(id);
+    const patientsList = persistedMedico.pacientes;
+    if (patientsList && patientsList.length > 0){
+      throw new BusinessLogicException(`The doctor has patients and cannot be deleted`, BusinessError.PRECONDITION_FAILED)
+    }
     await this.medicoRepository.remove(persistedMedico);
   }
 }
